third_party/WebKit/LayoutTests/fast/dom/HTMLImageElement/image-picture-nested-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS document.getElementById("canary").clientWidth is 75
PASS document.getElementById("canary").currentSrc is "resources/blue_rect.jpg?0"
PASS document.getElementById("picture_in_picture").clientWidth is 800
PASS document.getElementById("picture_in_picture").currentSrc is "../../hidpi/resources/image-set-2x.png?2"
PASS document.getElementById("picture_in_div_in_picture").clientWidth is 800
PASS document.getElementById("picture_in_div_in_picture").currentSrc is "../../hidpi/resources/image-set-2x.png?2"
PASS document.getElementById("picture_in_div_with_source_above").clientWidth is 800
PASS document.getElementById("picture_in_div_with_source_above").currentSrc is "../../hidpi/resources/image-set-2x.png?2"
PASS document.getElementById("picture_in_the_way").clientWidth is 800
PASS document.getElementById("picture_in_the_way").currentSrc is "../../hidpi/resources/image-set-2x.png?2"
PASS document.getElementById("picture_in_picture_1").clientWidth is 800
PASS document.getElementById("picture_in_picture_1").currentSrc is "../../hidpi/resources/image-set-2x.png?2"
PASS document.getElementById("picture_in_picture_2").clientWidth is 800
PASS document.getElementById("picture_in_picture_2").currentSrc is "../../hidpi/resources/image-set-2x.png?2"
PASS document.getElementById("picture_at_end").clientWidth is 75
PASS document.getElementById("picture_at_end").currentSrc is "resources/blue_rect.jpg?52"
PASS document.getElementById("picture_no_source").clientWidth is 75
PASS document.getElementById("picture_no_source").currentSrc is "resources/blue_rect.jpg?6"
 
 
 
     
